+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog
 ## 2.1.3
+- Added keybind modifiers
+- Added mouse wheel resizing
 - Fix sleep and thirst icon not showing
 - Fix hotkeys not being blocked when dialogs are open
-- Added keybind modifiers
-- 
 
 ## 2.1.2
 - Add mana/stamina checking for skills that require them